chore(actions): reoder actions directory - move all user actions into userActions.js
- move all group actions into groupActions.js

- move all message actions into messageActions.js

- export all actions from /client/src/actions/index.js
