--- client/src/actions/currentUserActions.js
@@ -1,8 +1,0 @@
-import { SET_CURRENT_USER } from '../../constant';
-
-const setCurrentUser = user => ({
-  type: SET_CURRENT_USER,
-  user
-});
-
-export default setCurrentUser;
