--- client/src/actions/flashMessageAction.js
@@ -1,8 +1,0 @@
-import * as types from '../../constant';
-
-const flashMessageAction = message => ({
-  type: types.NEW_MESSAGE,
-  message
-});
-
-export default flashMessageAction;
