+++ client/src/actions/index.js
@@ -1,0 +1,18 @@
+// User Actions
+export { userSignupRequest } from './userActions';
+export { userLoginRequest } from './userActions';
+export { logout } from './userActions';
+export { getAllUsers } from './userActions';
+export { searchUsers } from './userActions';
+export { forgotPassword } from './userActions';
+export { confirmPasswordResetRequest } from './userActions';
+export { setCurrentUser } from './userActions';
+
+// Group Actions
+export { createGroupRequest } from './groupActions';
+export { getUserGroups } from './groupActions';
+export { addUserRequest } from './groupActions';
+
+// Message Actions
+export { postMessageRequest } from './messageActions';
+export { getGroupMessages } from './messageActions';
