+++ client/src/components/Header/Navbar.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { NavLink } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import logout from '../../actions/logoutActions';
+import { logout } from '../../actions';
 import initialState from '../../initialState';
 //
 /**
